♻ combine comment header part
