+++ src/components/FeedView/Comments/CommentHeader/styles.js
@@ -1,11 +1,22 @@
 import styled from 'styled-components';
 
 export const HeaderContainer = styled.div`
+  position: relative;
   display: flex;
   align-items: center;
+  width: 100%;
+  padding-bottom: 15px;
+  .btnWrap {
+    right: 6px;
+    top: 22px;
+  }
+  span {
+    letter-spacing: -0.5px;
+  }
 `;
 
 export const UserInfo = styled.div`
+  margin-left: 11px;
   .name {
     font-size: 15.5px;
     padding-right: 6px;
@@ -34,5 +45,14 @@ export const UserInfo = styled.div`
 export const InfoBlock = styled.div`
   display: flex;
   align-items: center;
-  margin-left: 11px;
+`;
+
+export const DateWrap = styled.div`
+  position: absolute;
+  top: 0px;
+  right: 6px;
+  font-size: 11px;
+  color: ${({ theme }) => theme.color.black_2};
+  line-height: 28px;
+  margin-bottom: auto;
 `;
